stupid cast
git-svn-id: https://libmpplas.svn.sourceforge.net/svnroot/libmpplas/libmpplas/trunk@416 ea26bf2b-611e-0410-8ca6-ad9ecb09ac58

+++ src/sources/Trascendentes.cpp
@@ -253,7 +253,7 @@ namespace mpplas{
     int precisionVieja=0;
     if( prec ){
       precisionVieja = R::precision();
-      R::precision( (prec * Constants::LOG_2_10) +1 );
+      R::precision( (int)(prec * Constants::LOG_2_10) +1 );
 //      cout << (int)(prec * Constants::LOG_2_10) +1 << endl;
     }
     R unQuinto; unQuinto.hacerUno(); unQuinto /= 5;
